+++ cmd/root.go
@@ -66,7 +66,6 @@ func init() {
 		removeDirCmd,
 		runCmd,
 		exploreCmd,
-		mcpServeCmd,
 		pluginCmd,
 		wizardCmd,
 		bugReportCmd,